Small changes to remove unused (and now absent) dependencies, and added some missing code (segment_plot.py). Sim now runs, so long as PYTHONPATH=C:path/to/your/prt-sim/ is added to your envionmental variables. Start the program from main.py, and consider using the config file in tests/OneStation_OneVehicle to run.
example (from prt-sim/pyprt/sim dir):
python main.py --config ../tests/OneStation_OneVehicle/sim_config.cfg

Note that it's running a new controller, and the sim is in the process of having it's innards replumbed. As a result, it can't do NEAR as much as it could before, but that should change rapidly.