Added new message types to api to control passengers walking from one station to another. Support for the new message types is fully implemented on the Sim side, but is untested at the moment (not even a quick test.)
Changed globals.py back to common.py. Turns out that both globals and global are already taken in Python (as a function and a keyword, respectively).

Changed much of the implementation of the event.Passenger class. Has a more flexible setup for recording how much time is spent waiting, walking, and riding in vehicles. Previously, this was inaccurate if a passenger transfered.

Added some additional post sim summary statistics for passengers.

Added a new test case, which requires that passengers walk a short distance from one station to another in order to reach their destination. Currently, they fail.