Added API and implementation for passengers walking between stations. Fixed case where TrajectorySolver wasn't finding a solution (only one accel min/max hit). Separated test and default data from unit test folder. Updated gtf_unitrans_hw to be correct (originally had some NaN track lengths). Updated gtf_controller.PassengerManager to make passengers walk between stations when it's faster than waiting. Not tested. Need to add passenger walking speed as a commandline option of the controller.
Under the gtf_controller, Vehicles now plan a trajectory so that they stop at the beginning of the station trackSegment. When they reach the station offramp, they reserve a berth in the station and replan their trajectory to park in the berth. Replanning a complex trajectory while in motion has revealed some bugs. More on this at the end of this commit log.

Did some work on the setup.py script, realized that a proper easy_install just won't be possible, at least under windows. Too many dependencies which don't install well from easy_install. Currently planning to use pythonxy project to satisfy most dependencies (this includes the Enthought Tool Suite -- but I'm not totally sure yet if avoids the costs associated with the Enthought Python Distribution that come with non-educational use. I expect it does...). Need to write a full script to do it, though, since even the 'Full' install (~500 MB download!) doesn't install some of the necessary pieces by default. Alternatively, I should just bite the bullet and start working on a NSIS installer. MacOS and Linux aren't so bad, since Fortran and C/C++ compilers are easily available -- though even there, easy_install doesn't do a good job.

As an aside, I'm now using Python 2.6 (x86 not AMD64), since the dependencies support it now, and since python2.5 no longer has binary installers available from Python.org.

This commit doesn't work with the gtf_unitrans_hw test yet. It may not work with any test. I've decided to rewrite the sim's path/trajectory code to simplify and hopefully eradicate the head-ache inducing problems I've been having with it. As a bonus, it should be much more efficient too. Currently, splines are broken into Segments, and Segments are held by a Traversal (A Traversal is a container which holds all the Segments for a particular piece of track). Each segment starts at time 0, and the first Segment of every Traversal has a starting position of 0. Lots of shifting and discontinuities are required to maintain this structure, it requires very precise splitting of segments at a given position, and it doesn't gain me anything. It's complex, and doesn't need to be. Furthermore, I'm frequently finding the roots of a cubic polynomial, which is expensive.

The new implementation will wholly separate the concepts of a vehicle's Path (the itinerary of TrackSegments it travels over) and the vehicle's Trajectory (it's cubic spline). Each segment of the spline will be valid from it's start time to it's end time, as compared to being valid from t=0 to t=duration. This is exactly how cubic_spline.CubicSpline is already implemented, and I will be reusing and extending that class. I will still need to adjust the vehicle's position value to account for each vehicle's 'position' reseting to 0 whenever it reaches a new TrackSegment. Rather than manipulating the trajectory to do this, I will just keep a tally of the previous TrackSegment lengths, and subtract that from the current position. Python uses doubles (about 16 decimal digits accuracy) for all floating-point values, so even if a vehicle travels at 100 meters/sec for 86,400 seconds (24 hrs), it's position should still be accurate to ~8 places after the decimal -- more than plenty.