Largest change is in the Vehicle class, which was nearly entirely rewritten. It now utilizes the CubicSpline class that was developed for TrajectorySolver to hold the vehicle's current Trajectory. The vehicle's path (list of TrackSegments to visit, in order) is now completely separated from it's planned trajectory, which simplified the internals a great deal without losing any functionality.
Lots of various bugfixes. Though I won't make any guarantees about either stability or correctness yet, the project as a whole is in much better shape.

Small changes to the TrackBuilder side of the house. Fixed the problem with GTF data being dropped, and tweaked some settings for the GTF vehicle models.