Added a script that converts an a TrackBuilder save file into a Google Transit Feed shapes.txt file. It's pretty limited, and can only handle a single, closed loop of track at the moment. Making it more capable wouldn't be too difficult if the need arises.
Added the ability to abstractly refer to the current pyprt directory when writing config files. This allows the config files to only use relative paths, and thus be placed anywhere on the users system.

Small workaround added for a bug in TrackBuilder's GTF import. The layouts created from Transit Feeds no longer incorperate elevation data when calculating berth positions. This helps alleviate the problem with NaN's cropping up in the layout. An installer using the new code (ver 0.3.2) is on the website.