Fixes for prt controller. A key problem was that the controller was using out of date vehicle status info when calculating some trajectories. In most cases, this was inconsequential, but when doing a long run at full speed, the inaccuracy led to significant position error. The requirement for exact vehicle status info (vehicle pose) is a byproduct of the fire-and-forget nature of this controller. A more realistic controller would poll for status updates during the run and adjust the trajectory accordingly. The current fix is to always request a vehicle status update before calculating a trajectory.
The controller is still not robust enough for long or complex runs. The next main issue to fix is that the controller does not handle the case where a vehicle wants to enter a station, but the station's unload berths are all occupied (a NoBerthAvailableError is raised and is not caught). The fix will be to try reserving a berth before taking the station's offramp -- if no berth is available, stay on the main line and reroute the vehicle to the same station at a later time.