Small changes, but the sim can now run the a 60 vehicle tests under heavy passenger load for 2 (simulated) hours without problem.
Note that setting the FPS to 0 in the config file keeps the GUI, but disables the live visualization for a speed boost. This has been in for a long time, but is more useful as the simulations grow in complexity.

I hacked in some additional logging (now removed) for the position errors and found, to my relief, that they do not increase in magnitude as the sim progresses. In almost all cases the error is less than 1E-4 meters, but the sim was stumbling on a few cases where the error would be larger (up to 20 cm) during very short duration segments. I loosend the error thresholds in the asserts so as to only catch larger errors, or ones which accumulate.