Mostly TrackBuilder stuff. Finally redid the Select menu, so that I can find a particular TrackSegment, Station, or Vehicle. The whole concept of "selecting" something is still pretty incomplete, in that the currently selected item(s) are highlighted, but it doesn't DO anything. The improvements to the select menu were mostly because I needed them for debugging larger scenarios.
Small improvement to the prt_controller, in that empty vehicles aren't quite so restless to wander when there's no demand. There's still an issue with one test scenario (not all the test data is checked into the repository), wherein a vehicle doesn't slow to enter a station. Root cause still unknown.