TrackBuilder: Hopefully the last rewrite that will ever be required for the elevation requesting code. Rewrite was prompted by getting locked out of Google's elevation service for several hours -- the old version was just too spammy.
New version is much smarter about requesting elevations in big batches, with a 5 second pause between requests. The redundant requests are gone. Should be reasonably robust. It also grabs elevations at intervals along a track segment, and includes that info in the xml file, which was something I've wanted for energy calculations in the sim (just endpoint elevations was pretty coarse).

Also switched to using fixed lengths for the position, lat, and lng values in the xml file. Makes them a touch more readable, and gives a better indication of what the actual precision used is.

TODO:
 - Expose the elevation sampling interval in the user interface. Add to preferences file.
 - When parsing an xml file, add the elevations to the cache.
 - Improve the "View -> Grade" option to use the extra elevation data that's now available.
 - Add a "View -> Elevation" option to see absolute elevations using a color scale. (optional)
 - Status bar elevation indicator is disabled again. Either fix it or remove it.