New event type, SimNotifyVehicleSpeeding that is triggered when a vehicle exceeds either the vehicle's max speed (emergency, not comfort) or a track segment's speed limit.
The other new feature is I wrote my own quadratic and cubic root solvers and I am no longer relying on numpy.root() for finding the roots of polynomials. While the numpy library is generally excellent, numpy.root was not well suited for several reasons:
 1. It finds both imaginary and real roots. For the most part, I'm only interested in real roots.
 2. It wasn't precise. It wasn't uncommon for the roots it found to be off by 1E-5 or more. While still small errors, they were far, far larger than just rounding errors. This error is in comparison to Mathematica, and to my new code (which agreed with Mathematica).
 3. It was slow. For quadratics, my new code is about 100 times faster. For cubics, my code is about 25 times faster.

Why would anyone use numpy.root? It's capable of handling any order of polynomial, a feature I don't need.

One extra feature of my root solver is that I can specify that any inflection point that comes within threshold distance of the x-axis can be treated as though it were tangent to the x-axis. This nicely handles small errors in the coefficents leading to no root (inflection point doesn't quite reach the x-axis) or to two slightly separated roots (inflection point crosses the x-axis, instead of being tangent to it). In cases where the inflection point is further away from the x-axis (that is, not clearly intended to be tangent to it), there is no effect.

There is also a new command line option to profile the performance of the Sim thread, so as to find bottlenecks. Haven't played with this much yet, though, and it's not guaranteed to be working at this time.