Corresponds to TrackBuilder release 0.6.7
The Vehicle, Station, and TrackSegment classes now maintain references to their corresponding overlays (visual components). The above classes, and their corresponding overlay classes have all had a remove() member added which handles deleting the object from the scenario in an undoable fashion. The container classes, Vehicles,  Stations, and Tracks now have members which remove the object and it's associated overlay and removes it from the container (also undoable).

Deleting a TrackSegment which is a component of a station now informs the user that the TrackSegment and station must be deleted together. If the user clicks delete, the station and all of its TrackSegments, and any vehicles residing on them are all deleted (as one Undoable command).

This patch should fix the various and apparently sporadic problems with deleting pieces. One task left undone is that a TrackSegment that was 'split' when adding a station should really be 'mended' back together if the station is deleted. This occurs when the user Undoes the addition of the station, but not when it's deleted.