Missing info for the last few commits:
Highlights:
 - Added bb_freeze_setup.py script, which 'freezes' the application to a Windows executable.
 - Added inno_setup/inno_setup_script.iss which creates a 32-bit installer for Windows
 - Refreshed and updated all non-gtf scenarios
 - Testing scenarios are now stored in pyprt/scenarios rather than pyprt/data
 - All configuration files should now use a .ini extension, so that Windows knows to open them with a text editor.
 - Configuration files specify the controller type (PRT, GTF, or CUSTOM) and the arguments separately. This change makes it so that the path to the controller process or executable is no longer hardcoded in the config file. This change was necessary to support running from either a source tree checkout, or from an installed version.

This commit:
Removed all 'print' statements from the sim and controllers. When running as a non-console app (frozen, installed, gui_only=True) a print statement would cause the program to hang or crash. It may be possible to redirect via:
sys.stdout = open(os.devnull, 'w')
but that didn't work when I tried it. It may be that it needs to be done for every thread (or module?). At any rate, redirecting didn't work, but removing the print statements does.